return long

+++ ssl/ssl_sess.c
@@ -916,11 +916,11 @@ long SSL_SESSION_get_time(const SSL_SESSION *s)
     return (long)ossl_time_to_time_t(s->time);
 }
 
-long long SSL_SESSION_get_rtt(const SSL_SESSION *s) // might have to return a different data type. Or pass in a uint64_t* from nginx to fill in
+long SSL_SESSION_get_rtt(const SSL_SESSION *s) // might have to return a different data type. Or pass in a uint64_t* from nginx to fill in
 {
     if (s == NULL)
         return 0;
-    return ossl_time_to_time_t(s->rtt); // used to return (long).. // could be data loss here. Conversion from long long to long
+    return (long) ossl_time_to_time_t(s->rtt); // could be data loss here. Conversion from long long to long
 }
 
 long SSL_SESSION_set_time(SSL_SESSION *s, long t)
